python-extractor/requirements.txt
# FastAPI and server
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Browser automation
playwright==1.40.0

# ML models and image processing
transformers==4.35.2
torch==2.1.1
pillow==10.1.0
scikit-learn==1.3.2

# HTML parsing and text extraction
beautifulsoup4==4.12.2
lxml==4.9.3

# Async support
aiohttp==3.9.1
asyncio==3.4.3

# Optional: For better color extraction
opencv-python-headless==4.8.1.78
numpy==1.26.2
